requirements2.txt
dash[diskcache]
geopandas
dash_bootstrap_components
dash_daq
dash_extensions
pymysql
sshtunnel
Pillow
tqdm